moved class to directory
