+++ init_objects.py
@@ -1,12 +1,12 @@
-from XML_handling import MazeDAOXML
+from classes.XML_handling import MazeDAOXML
 mazeDAO = MazeDAOXML("saved_levels", "levels.xml")
-from Maze_class import Maze
+from classes.Maze_class import Maze
 maze = Maze(level=3)
-from Robot_flood_class import Robot
+from classes.Robot_flood_class import Robot
 robot = Robot()
 import tkinter as tk
 root = tk.Tk()
-from MazeView_class import MazeView
+from classes.MazeView_class import MazeView
 canvas = MazeView(root)
-from MazeApp_class import MazeApp
+from classes.MazeApp_class import MazeApp
 window = MazeApp(root)
