--- maze.xml
@@ -1,1 +1,0 @@
-<Maze><Row id="0"><Cell>1</Cell><Cell>2</Cell><Cell>3</Cell></Row><Row id="1"><Cell>4</Cell><Cell>5</Cell><Cell>6</Cell></Row><Row id="2"><Cell>7</Cell><Cell>8</Cell><Cell>9</Cell></Row></Maze>
